Change side of Essentia Meter
To place an Essentia Meter at the output of a Funnel, the hopper must face the opposite direction from where you want to position the Essentia Meter. This way, you can place a chest on the other side.
You can lock the hopper below the funnel by right-clicking the funnel with an empty hand (this will send a redstone signal to the hopper).
Fixed a bug: The label on the Essentia Meter was lost when the block was broken, but now it is preserved.